1-й семинар - правки после ревью

+++ basic/src/main/java/academy/backend/market_pulse/model/Bond.java
@@ -1,21 +1,23 @@
 package academy.backend.market_pulse.model;
 
+import java.math.BigDecimal;
+
 /**
  * Облигация.
  */
 public class Bond extends Instrument {
 
-    private final double couponRate;
+    private final BigDecimal couponRate;
     private final int maturityYear;
 
     public Bond(String ticker, String name, Currency currency,
-                double couponRate, int maturityYear) {
+                BigDecimal couponRate, int maturityYear) {
         super(ticker, name, currency);
         this.couponRate = couponRate;
         this.maturityYear = maturityYear;
     }
 
-    public double getCouponRate() {
+    public BigDecimal getCouponRate() {
         return couponRate;
     }
